+++ src/SmkingServiceProvider.php
@@ -10,6 +10,7 @@ use Illuminate\Support\ServiceProvider;
 use ReflectionClass;
 use ReflectionException;
 use Smking\Laravel\Console\CachePurgeCommand;
+use Smking\Laravel\Console\CircuitStatusCommand;
 use Smking\Laravel\Console\DoctorCommand;
 use Smking\Laravel\Http\Middleware\InjectAeo;
 use Smking\Laravel\View\Components\Aeo as AeoComponent;
@@ -58,6 +59,7 @@ class SmkingServiceProvider extends ServiceProvider
             $this->commands([
                 DoctorCommand::class,
                 CachePurgeCommand::class,
+                CircuitStatusCommand::class,
             ]);
         }
     }